Add PrettyLog usage for test
