+++ model/__init__.py
@@ -1,4 +1,5 @@
 from .unet_att import UnetAttention
+import pprint
 
 def build_model(cfg):
     print(f'loading model {cfg.name}')
@@ -11,3 +12,8 @@ def build_model(cfg):
     model = model.to(cfg.device)       
     return model
 
+class PrettyLog():
+    def __init__(self, obj):
+        self.obj = obj
+    def __repr__(self):
+        return pprint.pformat(self.obj)
